aiservice/requirements.txt
fastapi
pydantic
python-dotenv
uvicorn
langchain
google-generativeai
langchain-google-genai